Mutations and inputs
Create mutation to create a course
