+++ lib/mutations.js
@@ -1,0 +1,26 @@
+'use strict';
+
+const connectDb = require('./db');
+
+module.exports = {
+	createCourse: async (root, { input }) => {
+		const defaults = {
+			teacher: '',
+			topic: ''
+		};
+
+		const newCourse = Object.assign(defaults, input);
+		let db;
+		let course;
+
+		try {
+			db = await connectDb();
+			course = await db.collection('courses').insertOne(newCourse);
+			newCourse._id = course.insertedId;
+		} catch (error) {
+			console.error(error);
+		}
+
+		return newCourse;
+	}
+};
