+++ lib/schemas.graphql
@@ -12,3 +12,15 @@ type Query {
   "Devuelve un curso"
   getCourse(id: ID!): Course
 }
+
+input CourseInput {
+  title: String!
+  teacher: String
+  description: String!
+  topic: String
+}
+
+type Mutation {
+  "Crea un curso"
+  createCourse(input: CourseInput!): Course
+}
